requirements.txt
requests
pymysql==0.9.3
Django==2.2.7
djangorestframework==3.10.3
mysqlclient==1.4.6
markdown
coreapi
django-redis
pillow